requirements.txt
streamlit>=1.28.0
google-adk>=1.20.0
python-dotenv>=1.0.0
# Optional: for prescription image OCR (install Tesseract on system separately)
Pillow>=10.0.0
pytesseract>=0.3.10
